New functions for writing VTK and freesurfer surfs
Now it is possible to write 'legacy' VTK (plain ascii surface definition)
and freesurfer ascii meshes when the vertex and faces are well defined.

I checked visually orientation and consistency with paraview for VTK and
tkmedit for freesurfer ascii. There still remains a little offset with
freesurfer, but I'm not sure that it'll happen with real images.

For now, I close #60. Then re-open if new issues are found.