Fixed floating point exception (bug in #7)
Now, we have a first implementation of the anisotropic and inhomogeneous optimization.
I will chat with @zosso about how to initialize the A and B matrices with meaningful values.
In this commit, I just close #7. This means that I won't test the process here. If needed,
I will open a new issue to check the correctness of the implementation.