New approach to solve #77
It was too messy to have the ROIs and then apply the warping. New
approach is:

  1. Transform contour (actually, it is get the current position bc I cache it)
  2. Compute binary mask from it (at high resolution)
  3. Downsample with averaging to ref. grid -> should provide a fuzzy mask

This way, I expect to have nice weights for estimating the new parameters.

For computing energy, it is also necessary to compute the Jacobian map. For
re-estimation, I guess it shouldn't. But I will think around this a bit more.