Implemented parameters estimation routine
Now it is possible to re-estimate parameters whenever we want, so
close #70.

On the other hand, I realized that our approach is 2-class-only
segmentation. I opened issue #76 raising the concerns regarding
surfaces overlap and this stuff... but the problem is bigger than
this.

In the update formulas, the gradients are computed ONLY with the
*inside*/*outside* statistics. The problem comes when a region
is neighbouring with more than 1 other region. In this case,
our current implementation doesn't take it into account when computing
the *outside* energy.

I think it'd be enough to set inside/outside information on EACH vertex
when initializing because we do not want to change topology.