Theoretical demonstrations - TMI paper
A base for this request has been written. Close #67. This demonstration
has not been validated by @zosso yet, and it does not claim to be
complete.

But a base draft for the paper has been written. I will open a new
milestone that will gather all the issues regarding the details of
the paper. All discussions around the demonstration should go there.

The development of the paper will remain on branch paper/TMI