Hierarchical regions organization
Now contours are considered *layers*, in a way that they are
checked for consistency (#76).

Also, FreeSurferTest has been tested with several contours,
getting a coherent registration results and the implementation
of m_Modified has been changed to be able to close #85.

Anyways, the need for assigning the outer region to each vertex
is still open.