Added GTest
I add GTest to enable writing fast tests in the Google test way.
This should help avoid problems that never fix and/or changes that
affect other methods and then things stop running.

It is also good for when we want to publish the code.

Added a first base file (LevelSetsBaseTest.cxx) and an exemplary
CMakeLists.txt code.

Should help to close #76