Implementation of interpolator with kernel
Close #90. I've implemented an ITK compliant code, using the kernels
provided by ITK. I set BSpline kernel as default. I visually checked
forward (mesh to grid) interpolation and backward interpolation (from
grid to mesh).

The fwd interpolation seems ok. I've checked the ranges with a test
field of 3 control points. Intensity range in the field is very close
to that of the control points.

The bwd interpolation seems enough. For very wide kernels (currently
I've set a default kernel of 2mm), the result is not good. This is
the recovered values after performing forward and backward:

[-1.40729, 0.656734, -0.281458] vs. [-1.5, 0.7, -0.3]
[1.40721, 0.656697, 0.281441] vs. [1.5, 0.7, 0.3]
[-1.12583, 2.53312, 1.21965] vs. [-1.2, 2.7, 1.3]

Therefore, recovered points were close to the original ones... but
I do not know if this interpolation scheme is very reliable.

I'm going to test it on more complicated examples to see the behaviour.