New version of scripts
Improved T1 to dMRI transformation, as well as the surfaces
transformation. Now, the surfaces look the same when looking
at them in Slicer or Freeview (freesurfer). However, Paraview
still has an X-axis flip with origin error (it centers the
nifti volume?). Probably it is a bug the nifti plugin.

This commit relates #63 #53.

I will also need to check that #72 is not broken again.