Exposing new parameters in interface
In order that reporting all available information (#11), I've extended
the inputs and outputs in the command line interface (CLI). This, along
with the PR to git@github.com:nipy/nipype.git including the computation
of Dice and Jaccard fuzzy indexes, should be enough for getting a detailed
outcome of the registration process and automating the best parameters
search with nipype.