Reviewed region generation from meshes
There is no problem on this, and the mesh generated is consistent
with the resampling grid. Therefore, the input mesh should be
already well oriented. Close #91.

Nevertheless, it is advisable to rotate input reference image to
the orientation with identity direction matrix to avoid further
problems. This has been included in regseg Application.