Orientation of the vtk mesh
The orientation of the target image is not taken into account by
most of the visualization software (namely, I've tested freeview from
FreeSurfer and Slicer4).

The priors are now resampled to the reference image orientation,
in order to solve #104 and #105. If things are correctly carried
out, it should also affect #106.