Generated a full gold standard for IPMI Phantoms
I completely reproduced phantom 1 (sulcus model) and phantom 2.
Phantom 2 must be re-simulated again, so probably there are two
possible options:
1. Include some CSF in phantom 1 and ommit phantom 2
2. Improve Manu's phantom and also ommit phantom 2

Therefore, I don't think it is worth it to mantain phantom 2 with
the spheres. If we do, we need that @daducci generates again the
signal because he gave me a 30 directions volume (b0+29dwis) but
the bvec,bvals correspond to 31 directions (b0+30). And now
everything is messed up. FSL was working ok, but I don't know which
direction is removed to compute the tensor... Diffusion toolkit
failed fitting the tensor.

So, as phantom 1 works correctly, and phantom 2 does not require
further work until Alessandro is back, I close #107.