Refactored SparseMatrixTransform
Now I use OffGrid, OnGrid and Coefficient terms. OffGrid can be values
(displacements) or positions (of the mesh points). OnGrid can be values
(of the densified vector field) or positions (of the control points of the
bsplines). Finally, we have Coefficients of the BSpline or CoeffDerivatives
of the BSpline coefficients.