Memory allocation lowered
Initially I started using 15GiB when using the SparseMatrixTransform
to interpolate high-density fields. Right now, the same experiment
takes only 3.6GiB. It's still a lot, but it does not block computers
with more than that amount.

There's still one minor section of the code that should be revised,
right now the LU solver of sparse matrices is disabled. This must
be addressed before I can close the issue #142.